% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_account.R
\name{create_account}
\alias{create_account}
\title{Create an account for using Techtonique APIs}
\usage{
create_account(username, password)
}
\arguments{
\item{username}{a string; your username}

\item{password}{a string; your password}
}
\description{
Create an account for using Techtonique APIs
}
