% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Obtain forecasts from a csv file}
\usage{
get_forecast(
  file,
  token,
  method = c("theta", "mean", "rw", "prophet"),
  h = 5,
  level = 95,
  date_formatting = "original",
  start_training = NULL,
  n_training = NULL
)
}
\arguments{
\item{file}{a string; path to csv file containing time series data (examples:
https://github.com/Techtonique/datasets/tree/main/time_series/univariate)}

\item{token}{a string; authentication token (obtained with \code{forecastingAPI::get_token})}

\item{method}{a string; forecasting method (currently Theta, mean forecast, Random walk forecasting,
and FB Prophet forecasting)}

\item{h}{an integer; number of periods for forecasting}

\item{level}{an integer; confidence levels for prediction intervals}

\item{date_formatting}{a string; either "original" (yyyy-mm-dd) or "ms" (milliseconds)}

\item{start_training}{an integer; starting index for training window}

\item{n_training}{an integer; length of training window}
}
\value{
a list of "averages" (mean forecast) and "ranges" (lower and upper bounds of prediction interval)
}
\description{
Obtain forecasts from a csv file
}
