% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Obtain a token, for use in API requests}
\usage{
get_token(username, password)
}
\arguments{
\item{username}{a string; your username}

\item{password}{a string; your password}
}
\value{
a string; the token used in API requests (function \code{get_forecast} of this package)
}
\description{
Obtain a token, for use in API requests
}
